Enable the use of a home directory in the s4c
This adds a re-usable dedicated home directory that can be re-used
across container invocations.

The actual user-home isn't used here as this home tends to be used for
build tools installed in the home directory that depend on the target
distribution, so we avoid the host and container from conflicting in the
versions of these tools they wish installed.

Signed-off-by: Curtis Millar <curtis@curtism.me>